+++ com.unity.formats.alembic/Editor/Exporter/AlembicExporterEditor.cs
@@ -12,12 +12,12 @@ namespace UnityEditor.Formats.Alembic.Exporter
     [CustomEditor(typeof(AlembicExporter))]
     class AlembicExporterEditor : Editor
     {
-        bool m_foldCaptureComponents;
-        bool m_foldMeshComponents;
+        SavedBool m_foldCaptureComponents;
+        SavedBool m_foldMeshComponents;
 
         internal static bool DrawSettings(SerializedObject so,
             AlembicRecorderSettings settings,
-            string pathSettings, ref bool foldCaptureComponents, ref bool foldMeshComponents, bool recorder)
+            string pathSettings, ref SavedBool foldCaptureComponents, ref SavedBool foldMeshComponents, bool recorder)
         {
             bool dirty = false;
             if (!recorder)
@@ -63,25 +63,6 @@ namespace UnityEditor.Formats.Alembic.Exporter
             // alembic settings
             EditorGUILayout.LabelField("Alembic Settings", EditorStyles.boldLabel);
             {
-#if !UNITY_EDITOR_LINUX
-                var archiveProp = so.FindProperty(pathSettings + "conf.archiveType");
-                using (var hScope = new EditorGUILayout.HorizontalScope())
-                {
-                    using (var propertyScope = new EditorGUI.PropertyScope(hScope.rect,
-                        new GUIContent(archiveProp.displayName), archiveProp))
-                    {
-                        using (var changeScope = new EditorGUI.ChangeCheckScope())
-                        {
-                            var val = EditorGUILayout.EnumPopup(propertyScope.content, (ArchiveType)archiveProp.intValue, null,
-                                true);
-                            if (changeScope.changed)
-                            {
-                                archiveProp.intValue = (int)(ArchiveType)val;
-                            }
-                        }
-                    }
-                }
-#endif
                 EditorGUILayout.PropertyField(so.FindProperty(pathSettings + "conf.xformType"));
                 if (!recorder)
                 {
@@ -119,7 +100,7 @@ namespace UnityEditor.Formats.Alembic.Exporter
             EditorGUILayout.PropertyField(so.FindProperty(pathSettings + "assumeNonSkinnedMeshesAreConstant"), new GUIContent("Static MeshRenderers"));
             GUILayout.Space(5);
 
-            foldCaptureComponents = EditorGUILayout.Foldout(foldCaptureComponents, "Capture Components");
+            foldCaptureComponents.value = EditorGUILayout.Foldout(foldCaptureComponents, "Capture Components");
             if (foldCaptureComponents)
             {
                 EditorGUI.indentLevel++;
@@ -129,7 +110,8 @@ namespace UnityEditor.Formats.Alembic.Exporter
                 EditorGUI.indentLevel--;
             }
 
-            foldMeshComponents = EditorGUILayout.Foldout(foldMeshComponents, "Mesh Components");
+
+            foldMeshComponents.value = EditorGUILayout.Foldout(foldMeshComponents, "Mesh Components");
             if (foldMeshComponents)
             {
                 EditorGUI.indentLevel++;
@@ -142,6 +124,12 @@ namespace UnityEditor.Formats.Alembic.Exporter
             }
 
             return dirty;
+        }
+
+        public void OnEnable()
+        {
+            m_foldCaptureComponents = new SavedBool($"{target.GetType()}.m_foldCaptureComponents", false);
+            m_foldMeshComponents = new SavedBool($"{target.GetType()}.m_foldMeshComponents", false);
         }
 
         public override void OnInspectorGUI()
@@ -199,6 +187,44 @@ namespace UnityEditor.Formats.Alembic.Exporter
                         t.OneShot();
                 }
             }
+        }
+    }
+
+    class SavedBool
+    {
+        bool m_Value;
+        string m_Name;
+        bool m_Loaded;
+        public SavedBool(string name, bool value)
+        {
+            m_Name = name;
+            m_Loaded = false;
+            m_Value = value;
+        }
+
+        private void Load()
+        {
+            if (m_Loaded)
+                return;
+            m_Loaded = true;
+            m_Value = EditorPrefs.GetBool(m_Name, m_Value);
+        }
+
+        public bool value
+        {
+            get { Load(); return m_Value; }
+            set
+            {
+                Load();
+                if (m_Value == value)
+                    return;
+                m_Value = value;
+                EditorPrefs.SetBool(m_Name, value);
+            }
+        }
+        public static implicit operator bool(SavedBool s)
+        {
+            return s.value;
         }
     }
 }