+++ com.unity.formats.alembic/Editor/Importer/AlembicBuildPostProcess.cs
@@ -7,6 +7,7 @@ using System.Text;
 using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
 using UnityEditor.Callbacks;
+using UnityEngine;
 using UnityEngine.Formats.Alembic.Importer;
 using UnityEngine.SceneManagement;
 using Object = UnityEngine.Object;
@@ -16,11 +17,20 @@ namespace UnityEditor.Formats.Alembic.Importer
     static class AlembicBuildPostProcess
     {
         internal static readonly List<KeyValuePair<string, string>> FilesToCopy = new List<KeyValuePair<string, string>>();
+        internal static bool HaveAlembicInstances = false;
         [PostProcessBuild]
         public static void OnPostProcessBuild(BuildTarget target, string pathToBuiltProject)
         {
             if (!TargetIsSupported(target))
             {
+                if (HaveAlembicInstances)
+                {
+                    Debug.LogWarning(
+                        "Alembic only supports the following build targets: Windows 64-bit, MacOS X, Linux 64-bit or Stadia");
+                }
+
+                HaveAlembicInstances = false;
+
                 return;
             }
 
@@ -39,7 +49,7 @@ namespace UnityEditor.Formats.Alembic.Importer
 
         public static bool TargetIsSupported(BuildTarget target)
         {
-            return target == BuildTarget.StandaloneOSX || target == BuildTarget.StandaloneWindows64 || target == BuildTarget.StandaloneLinux64;
+            return target == BuildTarget.StandaloneOSX || target == BuildTarget.StandaloneWindows64 || target == BuildTarget.StandaloneLinux64 || target == BuildTarget.Stadia;
         }
     }
 
@@ -54,6 +64,8 @@ namespace UnityEditor.Formats.Alembic.Importer
         {
             if (report == null || !AlembicBuildPostProcess.TargetIsSupported(report.summary.platform))
             {
+                AlembicBuildPostProcess.HaveAlembicInstances |= scene.GetRootGameObjects()
+                    .SelectMany(root => root.GetComponentsInChildren<AlembicStreamPlayer>(true)).Any();
                 return;
             }
 
@@ -87,6 +99,8 @@ namespace UnityEditor.Formats.Alembic.Importer
             {
                 case BuildTarget.StandaloneOSX:
                     return Path.Combine(summary.outputPath, "Contents/Resources/Data/StreamingAssets");
+                case BuildTarget.Stadia:
+                    return Path.Combine(summary.outputPath, "layout/files/Data/StreamingAssets");
                 case BuildTarget.StandaloneLinux64:
                 case BuildTarget.StandaloneWindows64:
                     var name = Path.ChangeExtension(summary.outputPath, null);